day91_colorExtractor/requirements.txt
click==8.0.1
colorama==0.4.4
convcolors==1.0.0
extcolors==1.0.0
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
webcolors==1.11.1
Werkzeug==2.0.1
